requirements.txt
blobfile==2.0.0
numpy==1.23.4
Pillow==9.5.0
scikit_image==0.19.3
tensorflow==2.10.0
torch==2.0.0
tqdm==4.64.1
